Create DSpaceRepository as parent with gen.purpose methods; add "this."
