+++ src/main/java/bwfdm/replaydh/workflow/export/dspace/DSpaceRepository.java
@@ -1,0 +1,179 @@
+/*
+ * Unless expressly otherwise stated, code from this project is licensed under the MIT license [https://opensource.org/licenses/MIT].
+ * 
+ * Copyright (c) <2018> <Markus Gärtner, Volodymyr Kushnarenko, Florian Fritze, Sibylle Hermann and Uli Hahn>
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), 
+ * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, 
+ * and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, 
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A 
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT 
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF 
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH 
+ * THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package bwfdm.replaydh.workflow.export.dspace;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.swordapp.client.AuthCredentials;
+import org.swordapp.client.ProtocolViolationException;
+import org.swordapp.client.SWORDClient;
+import org.swordapp.client.SWORDClientException;
+import org.swordapp.client.SWORDCollection;
+import org.swordapp.client.SWORDWorkspace;
+import org.swordapp.client.ServiceDocument;
+import org.swordapp.client.UriRegistry;
+
+/**
+ * The class consists implementation of some general purpose DSpace-specific methods.
+ * It should be used as a parent for further child-classes as e.g. DSpace_v6, Dspace_v5.
+ * 
+ * @author Volodymyr Kushnarenko
+ *
+ */
+public abstract class DSpaceRepository implements PublicationRepository{
+
+	private static final Logger log = LoggerFactory.getLogger(DSpaceRepository.class);
+	
+	// Header constants
+	public static final String APPLICATION_JSON = "application/json";
+	public static final String CONTENT_TYPE_HEADER = "Content-Type";
+	public static final String ACCEPT_HEADER = "Accept";
+	
+	
+	/*
+	 * -------------------------------
+	 * General purpose methods
+	 * -------------------------------
+	 */
+	
+	
+	/**
+	 * Get new authentication credentials. 
+	 * <p> To disactivate "on-behalf-of" option please use the same string for "adminUser" and "userLogin".
+	 * <p> If "adminUser" and "userLogin" are different, "on-behalf-of" option will be used.
+	 * 
+	 * @param adminUser
+	 * @param adminPassword
+	 * @param userLogin
+	 * @return
+	 */
+	public static AuthCredentials getNewAuthCredentials(String adminUser, char[] adminPassword, String userLogin) {
+		
+		if(adminUser.equals(userLogin)) {
+			return new AuthCredentials(userLogin, String.valueOf(adminPassword)); // without "on-behalf-of"
+		} else {
+			return new AuthCredentials(adminUser, String.valueOf(adminPassword), userLogin); // with "on-behalf-of"
+		}
+	}
+	
+	
+	/**
+	 * Get service document via SWORD v2
+	 * 
+	 * @param swordClient
+	 * @param serviceDocumentURL
+	 * @param authCredentials
+	 * @return ServiceDocument or null in case of error/exception
+	 */
+	public static ServiceDocument getServiceDocument(SWORDClient swordClient, String serviceDocumentURL, AuthCredentials authCredentials) {
+		ServiceDocument serviceDocument = null;
+		try {
+			serviceDocument = swordClient.getServiceDocument(serviceDocumentURL, authCredentials);
+		} catch (SWORDClientException | ProtocolViolationException e) {
+			log.error("Exception by accessing service document: " + e.getClass().getSimpleName() + ": " + e.getMessage());
+			return null;
+		}
+		return serviceDocument;
+	}
+	
+	
+	/**
+	 * Get available collections via SWORD v2
+	 * 
+	 * @return Map<String, String> where key=URL, value=Title
+	 */
+	public static Map<String, String> getAvailableCollectionsViaSWORD(ServiceDocument serviceDocument){
+		Map<String, String> collections = new HashMap<String, String>();
+		
+		if(serviceDocument != null) {
+			for(SWORDWorkspace workspace : serviceDocument.getWorkspaces()) {
+				for (SWORDCollection collection : workspace.getCollections()) {
+					// key = full URL, value = Title
+					collections.put(collection.getHref().toString(), collection.getTitle());
+				}
+			}
+		}
+		return collections;
+	}
+	
+	
+	/**
+	 * Get a file extension (without a dot) from the file name (e.g. "txt", "zip", ...)
+	 * @param fileName
+	 * @return
+	 */
+	public static String getFileExtension(String fileName) {	
+		String extension = "";
+		int i = fileName.lastIndexOf('.');
+		if(i>0) {
+			extension = fileName.substring(i+1);
+		}
+		return extension;		
+	}
+	
+	
+	/**
+	 * Get package format basing on the file name.
+	 * E.g. {@link UriRegistry.PACKAGE_SIMPLE_ZIP} {@link UriRegistry.PACKAGE_BINARY}
+	 * @param fileName
+	 * @return 
+	 */
+	public static String getPackageFormat(String fileName) {
+		String extension = getFileExtension(fileName);
+		
+		if(extension.toLowerCase().equals("zip")) {
+			return UriRegistry.PACKAGE_SIMPLE_ZIP;
+		}
+		return UriRegistry.PACKAGE_BINARY;
+	}
+		
+	
+	/*
+	 * -------------
+	 * Extra classes
+	 * -------------
+	 */
+	
+	
+	public static enum SwordRequestType {	
+		DEPOSIT("DEPOSIT"), //"POST" request
+		REPLACE("REPLACE"), //"PUT" request
+		DELETE("DELETE")	//reserved for the future
+		;
+		
+		private final String label;
+		
+		private SwordRequestType(String label) {
+			this.label = label;
+		}
+		
+		public String getLabel() {
+			return label;
+		}
+		
+		@Override
+		public String toString() {
+			return label;
+		}
+	}
+		
+}
